Much better handling of near_* actions.
Action handlers now need to deal with all cases
(ie, whether players nearby can see the character
performing the action) which makes the handlers
more complicated, but also gives better control
and better messages.

Actions are now passed in to containers in the
room (recursively) for possible handling.

Added DetectInvisRing to test out the new handlers.
Item allows character to detect invisible items
and characters in the room, but not to actually
see them.