Handle mobs in destroyed rooms better. (issue 23).
Teleport the mobs home when the room is destroyed, or if
the room destroyed is their home, destroy the mob also.

If the room destroyed is a mob's home, but the mob is not
home when it is destroyed, the mob's new home will be the
default home (void by default).