More unicode cleanup for utf-8 compatibility.
Should fix issue 30 and issue 37.

Note that you may need to create a fresh database due to
stray uncoverted characters. Make a backup first! There
is probably a way to upgrade a database that has problem
characters but I have not looked at it.

Please test and report and problems.